feat(ui): move model detailed info to a modal
